requirements.txt
flask==3.0.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
beautifulsoup4==4.12.2
requests==2.31.0
python-whois==0.8.0
tldextract==5.1.1
